+ telegram ID is now a uint64

+++ Dota2Cheat/UI/Pages/MainMenu.h
@@ -78,7 +78,7 @@ namespace Pages {
 					ImGui::SliderInt("Delay", &Config::AutoAccept::Delay, 0, 6);
 					ImGui::Checkbox("Send telegram notifications", &Config::AutoAccept::SendTelegramNotifications);
 					if (Config::AutoAccept::SendTelegramNotifications) {
-						ImGui::InputInt("Telegram ID", &Config::API::TelegramID);
+						ImGui::InputUInt64("Telegram ID", &Config::API::TelegramID);
 						ImGui::SameLine(); HelpMarker("First, start the bot at t.me/dotacheatnotifybot\nThen get your ID at t.me/getmyid_bot");
 					}
